// packages/shared/src/enums/msp.enum.ts
export enum MSPOrganizationType {
	ORGANIZATION_STAFFING_OFFICE = "ORGANIZATION_STAFFING_OFFICE",
	CORPORATE_OFFICE = "CORPORATE_OFFICE",
	BRANCH_OFFICE = "BRANCH_OFFICE",
	REMOTE_OFFICE = "REMOTE_OFFICE",
}

export enum OrganizationTimezone {
	EASTERN = "EASTERN",
	CENTRAL = "CENTRAL",
	MOUNTAIN = "MOUNTAIN",
	PACIFIC = "PACIFIC",
}